Refactor MaintainMachineImages to also update shoot
We want to test the end result rather than the implementation details: If a shoot's worker pools get updated with a new image or not. In order to do this, we moved all logic dealing with selecting a MachineImageVersion and updating the shoot in a new method. This allows to make a method like selectUpdatedMachineImages private again, as it is an implementation detail only.

Co-authored-by: Marco Voelz <marco.voelz@sap.com>